requirements.txt
alembic==0.9.2
appdirs==1.4.3
click==6.7
Flask==0.12.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
nose==1.3.7
packaging==16.8
psycopg2==2.6.1
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.10
Werkzeug==0.12.2
